[#556] The rest of VP functions moved to separate file
