+++ plugins/versionpress/versionpress.php
@@ -843,19 +843,6 @@ function vp_ajax_hide_vp_welcome_panel() {
 add_action('wp_ajax_vp_show_undo_confirm', 'vp_show_undo_confirm');
 
 //----------------------------------
-// Private functions
-//----------------------------------
-
-function vp_enable_maintenance() {
-    $maintenance_string = '<?php define("VP_MAINTENANCE", true); $upgrading = ' . time() . '; ?>';
-    file_put_contents(ABSPATH . '/.maintenance', $maintenance_string);
-}
-
-function vp_disable_maintenance() {
-    FileSystem::remove(ABSPATH . '/.maintenance');
-}
-
-//----------------------------------
 // CSS & JS
 //----------------------------------
